refactor: divide search query handle logic as a custom hook
